Add replaceNode parameter test file
