+++ test/browser/render.test.js
@@ -6,7 +6,8 @@ import {
 	getMixedArray,
 	mixedArrayHTML,
 	serializeHtml,
-	supportsDataList
+	supportsDataList,
+	sortAttributes
 } from '../_util/helpers';
 import { clearLog, getLog, logCall } from '../_util/logCall';
 
@@ -18,20 +19,6 @@ function getAttributes(node) {
 		attrs[node.attributes[i].name] = node.attributes[i].value;
 	}
 	return attrs;
-}
-
-// hacky normalization of attribute order across browsers.
-function sortAttributes(html) {
-	return html.replace(
-		/<([a-z0-9-]+)((?:\s[a-z0-9:_.-]+=".*?")+)((?:\s*\/)?>)/gi,
-		(s, pre, attrs, after) => {
-			let list = attrs
-				.match(/\s[a-z0-9:_.-]+=".*?"/gi)
-				.sort((a, b) => (a > b ? 1 : -1));
-			if (~after.indexOf('/')) after = '></' + pre + '>';
-			return '<' + pre + list.join('') + after;
-		}
-	);
 }
 
 describe('render()', () => {
@@ -885,52 +872,6 @@ describe('render()', () => {
 		expect(scratch.textContent).to.equal('01');
 	});
 
-	it('should call unmount when working with replaceNode', () => {
-		const mountSpy = sinon.spy();
-		const unmountSpy = sinon.spy();
-		class MyComponent extends Component {
-			componentDidMount() {
-				mountSpy();
-			}
-			componentWillUnmount() {
-				unmountSpy();
-			}
-			render() {
-				return <div>My Component</div>;
-			}
-		}
-
-		const container = document.createElement('div');
-		scratch.appendChild(container);
-		render(<MyComponent />, scratch, container);
-		expect(mountSpy).to.be.calledOnce;
-
-		render(<div>Not my component</div>, document.body, container);
-		expect(unmountSpy).to.be.calledOnce;
-	});
-
-	it('should double replace', () => {
-		const container = document.createElement('div');
-		scratch.appendChild(container);
-		render(<div>Hello</div>, scratch, scratch.firstElementChild);
-		expect(scratch.innerHTML).to.equal('<div>Hello</div>');
-
-		render(<div>Hello</div>, scratch, scratch.firstElementChild);
-		expect(scratch.innerHTML).to.equal('<div>Hello</div>');
-	});
-
-	it('should replaceNode after rendering', () => {
-		function App({ i }) {
-			return <p>{i}</p>;
-		}
-
-		render(<App i={2} />, scratch);
-		expect(scratch.innerHTML).to.equal('<p>2</p>');
-
-		render(<App i={3} />, scratch, scratch.firstChild);
-		expect(scratch.innerHTML).to.equal('<p>3</p>');
-	});
-
 	it('should not cause infinite loop with referentially equal props', () => {
 		let i = 0;
 		let prevDiff = options._diff;
@@ -952,171 +893,6 @@ describe('render()', () => {
 		render(<App>10</App>, scratch);
 		expect(scratch.textContent).to.equal('10');
 		options._diff = prevDiff;
-	});
-
-	describe('subsequent replaces', () => {
-		it("shouldn't remove elements", () => {
-			const placeholder = document.createElement('div');
-			scratch.appendChild(placeholder);
-			const App = () => (
-				<div>
-					New content
-					<button>Update</button>
-				</div>
-			);
-
-			render(<App />, scratch, placeholder);
-			expect(scratch.innerHTML).to.equal(
-				'<div>New content<button>Update</button></div>'
-			);
-
-			render(<App />, scratch, placeholder);
-			expect(scratch.innerHTML).to.equal(
-				'<div>New content<button>Update</button></div>'
-			);
-		});
-
-		it('should remove redundant elements', () => {
-			const placeholder = document.createElement('div');
-			scratch.appendChild(placeholder);
-			const App = () => (
-				<div>
-					New content
-					<button>Update</button>
-				</div>
-			);
-
-			render(<App />, scratch, placeholder);
-			expect(scratch.innerHTML).to.equal(
-				'<div>New content<button>Update</button></div>'
-			);
-
-			placeholder.appendChild(document.createElement('span'));
-			expect(scratch.innerHTML).to.equal(
-				'<div>New content<button>Update</button><span></span></div>'
-			);
-
-			render(<App />, scratch, placeholder);
-			expect(scratch.innerHTML).to.equal(
-				'<div>New content<button>Update</button></div>'
-			);
-		});
-	});
-
-	describe('replaceNode parameter', () => {
-		function appendChildToScratch(id) {
-			const child = document.createElement('div');
-			child.id = id;
-			scratch.appendChild(child);
-		}
-
-		beforeEach(() => {
-			['a', 'b', 'c'].forEach(id => appendChildToScratch(id));
-		});
-
-		it('should use replaceNode as render root and not inject into it', () => {
-			const childA = scratch.querySelector('#a');
-			render(<div id="a">contents</div>, scratch, childA);
-			expect(scratch.querySelector('#a')).to.equalNode(childA);
-			expect(childA.innerHTML).to.equal('contents');
-		});
-
-		it('should not remove siblings of replaceNode', () => {
-			const childA = scratch.querySelector('#a');
-			render(<div id="a" />, scratch, childA);
-			expect(scratch.innerHTML).to.equal(
-				'<div id="a"></div><div id="b"></div><div id="c"></div>'
-			);
-		});
-
-		it('should notice prop changes on replaceNode', () => {
-			const childA = scratch.querySelector('#a');
-			render(<div id="a" className="b" />, scratch, childA);
-			expect(sortAttributes(String(scratch.innerHTML))).to.equal(
-				sortAttributes(
-					'<div id="a" class="b"></div><div id="b"></div><div id="c"></div>'
-				)
-			);
-		});
-
-		it('should unmount existing components', () => {
-			const newScratch = setupScratch();
-			const unmount = sinon.spy();
-			const mount = sinon.spy();
-			class App extends Component {
-				componentDidMount() {
-					mount();
-				}
-
-				componentWillUnmount() {
-					unmount();
-				}
-
-				render() {
-					return <div>App</div>;
-				}
-			}
-			render(
-				<div id="a">
-					<App />
-				</div>,
-				newScratch
-			);
-			expect(newScratch.innerHTML).to.equal('<div id="a"><div>App</div></div>');
-			expect(mount).to.be.calledOnce;
-			render(<div id="a">new</div>, newScratch, newScratch.querySelector('#a'));
-			expect(newScratch.innerHTML).to.equal('<div id="a">new</div>');
-			expect(unmount).to.be.calledOnce;
-
-			newScratch.parentNode.removeChild(newScratch);
-		});
-
-		it('should unmount existing components in prerendered HTML', () => {
-			const newScratch = setupScratch();
-			const unmount = sinon.spy();
-			const mount = sinon.spy();
-			class App extends Component {
-				componentDidMount() {
-					mount();
-				}
-
-				componentWillUnmount() {
-					unmount();
-				}
-
-				render() {
-					return <span>App</span>;
-				}
-			}
-
-			newScratch.innerHTML = `<div id="child"></div>`;
-
-			const childContainer = newScratch.querySelector('#child');
-
-			render(<App />, childContainer);
-			expect(serializeHtml(childContainer)).to.equal('<span>App</span>');
-			expect(mount).to.be.calledOnce;
-			render(<div />, newScratch, newScratch.firstElementChild);
-			expect(serializeHtml(newScratch)).to.equal('<div id=""></div>');
-			expect(unmount).to.be.calledOnce;
-
-			newScratch.parentNode.removeChild(newScratch);
-		});
-
-		it('should render multiple render roots in one parentDom', () => {
-			const childA = scratch.querySelector('#a');
-			const childB = scratch.querySelector('#b');
-			const childC = scratch.querySelector('#c');
-			const expectedA = '<div id="a">childA</div>';
-			const expectedB = '<div id="b">childB</div>';
-			const expectedC = '<div id="c">childC</div>';
-			render(<div id="a">childA</div>, scratch, childA);
-			render(<div id="b">childB</div>, scratch, childB);
-			render(<div id="c">childC</div>, scratch, childC);
-			expect(scratch.innerHTML).to.equal(
-				`${expectedA}${expectedB}${expectedC}`
-			);
-		});
 	});
 
 	it('should not call options.debounceRendering unnecessarily', () => {
